Pull the glossary summary into a front matter field
The current summary in the body text uses the glossary_tooltip shortcode, which currently uses the summary provided by the content.

There are, however, some ininite recursions here. Hugo handles is decently (creates some WARNINGs), but this should be avoided.

So, this commit makes the tooltip shortcode use the front matter summary instead, that doesn't have this problem.